refactor: validateUserInfo 인자를 객체로 받아오기
